+++ apps/server/test/modules/learnroom/dashboard.e2e-spec.ts
@@ -55,7 +55,7 @@ describe('Dashboard Controller (e2e)', () => {
 				grid: [
 					{
 						pos: { x: 1, y: 3 },
-						gridElement: GridElement.FromSingleReference(
+						gridElement: GridElement.FromPersistedReference(
 							new ObjectId().toString(),
 							new DefaultGridReference(new ObjectId().toString(), 'Mathe')
 						),
@@ -77,14 +77,14 @@ describe('Dashboard Controller (e2e)', () => {
 				grid: [
 					{
 						pos: { x: 1, y: 3 },
-						gridElement: GridElement.FromSingleReference(
+						gridElement: GridElement.FromPersistedReference(
 							new ObjectId().toString(),
 							new DefaultGridReference(new ObjectId().toString(), 'Quantumphysics')
 						),
 					},
 					{
 						pos: { x: 2, y: 2 },
-						gridElement: GridElement.FromSingleReference(
+						gridElement: GridElement.FromPersistedReference(
 							new ObjectId().toString(),
 							new DefaultGridReference(new ObjectId().toString(), 'Astrophysics')
 						),
@@ -109,14 +109,14 @@ describe('Dashboard Controller (e2e)', () => {
 				grid: [
 					{
 						pos: { x: 2, y: 2 },
-						gridElement: GridElement.FromSingleReference(
+						gridElement: GridElement.FromPersistedReference(
 							new ObjectId().toString(),
 							new DefaultGridReference(new ObjectId().toString(), 'mannequinization')
 						),
 					},
 					{
 						pos: { x: 3, y: 3 },
-						gridElement: GridElement.FromReferenceGroup(new ObjectId().toString(), [
+						gridElement: GridElement.FromPersistedGroup(new ObjectId().toString(), [
 							new DefaultGridReference(new ObjectId().toString(), 'Perspective Drawing'),
 							new DefaultGridReference(new ObjectId().toString(), 'Shape Manipulation'),
 						]),
@@ -136,12 +136,36 @@ describe('Dashboard Controller (e2e)', () => {
 			expect(body.gridElements[0]).toHaveProperty('groupElements');
 		});
 
+		it('should remove element from group', async () => {
+			const dashboard = new DashboardEntity(new ObjectId().toString(), {
+				grid: [
+					{
+						pos: { x: 3, y: 3 },
+						gridElement: GridElement.FromPersistedGroup(new ObjectId().toString(), [
+							new DefaultGridReference(new ObjectId().toString(), 'Perspective Drawing'),
+							new DefaultGridReference(new ObjectId().toString(), 'Shape Manipulation'),
+						]),
+					},
+				],
+			});
+			await dashboardRepo.persistAndFlush(dashboard);
+
+			const params = {
+				from: { x: 3, y: 3, groupIndex: 0 },
+				to: { x: 2, y: 3 },
+			};
+			const response = await request(app.getHttpServer()).patch(`/dashboard/${dashboard.id}/moveElement`).send(params);
+			expect(response.status).toEqual(200);
+			const body = response.body as DashboardResponse;
+			expect(body.gridElements.length).toEqual(2);
+		});
+
 		it('should fail with incomplete input', async () => {
 			const dashboard = new DashboardEntity(new ObjectId().toString(), {
 				grid: [
 					{
 						pos: { x: 1, y: 3 },
-						gridElement: GridElement.FromSingleReference(
+						gridElement: GridElement.FromPersistedReference(
 							new ObjectId().toString(),
 							new DefaultGridReference(new ObjectId().toString(), 'Mathe')
 						),